+++ Assets/Scripts/MainRoom/SpawnCharacter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 65227af2de599c844a4e493520615c78
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
